Update code to my current version
This now displays the number of devices found during a poll session, only if a potential skimmer was found.  This does NOT do a handshake yet.  I'm still working on that part, and learning Python for the first time in my life, so I have to reverse engineer the language a little.